moving DataSetService from dao to dataset package
