+++ api/src/main/java/org/pmiops/workbench/dataset/DataSetServiceImpl.java
@@ -1,4 +1,4 @@
-package org.pmiops.workbench.db.dao;
+package org.pmiops.workbench.dataset;
 
 import static com.google.cloud.bigquery.StandardSQLTypeName.ARRAY;
 import static org.pmiops.workbench.model.PrePackagedConceptSetEnum.SURVEY;
@@ -35,7 +35,9 @@ import org.pmiops.workbench.cohortbuilder.CohortQueryBuilder;
 import org.pmiops.workbench.cohortbuilder.ParticipantCriteria;
 import org.pmiops.workbench.config.CdrBigQuerySchemaConfig;
 import org.pmiops.workbench.config.CdrBigQuerySchemaConfigService;
-import org.pmiops.workbench.dataset.BigQueryDataSetTableInfo;
+import org.pmiops.workbench.db.dao.CohortDao;
+import org.pmiops.workbench.db.dao.ConceptSetDao;
+import org.pmiops.workbench.db.dao.DataSetDao;
 import org.pmiops.workbench.db.model.DbCohort;
 import org.pmiops.workbench.db.model.DbConceptSet;
 import org.pmiops.workbench.db.model.DbDataset;
@@ -165,11 +167,11 @@ public class DataSetServiceImpl implements DataSetService, GaugeDataCollector {
       this.namedParameterValues = namedParameterValues;
     }
 
-    String getQuery() {
+    public String getQuery() {
       return query;
     }
 
-    Map<String, QueryParameterValue> getNamedParameterValues() {
+    public Map<String, QueryParameterValue> getNamedParameterValues() {
       return namedParameterValues;
     }
   }
