+++ api/src/test/java/org/pmiops/workbench/db/dao/DataSetServiceTest.java
@@ -36,7 +36,8 @@ import org.pmiops.workbench.api.BigQueryService;
 import org.pmiops.workbench.cdr.ConceptBigQueryService;
 import org.pmiops.workbench.cohortbuilder.CohortQueryBuilder;
 import org.pmiops.workbench.config.CdrBigQuerySchemaConfigService;
-import org.pmiops.workbench.db.dao.DataSetServiceImpl.QueryAndParameters;
+import org.pmiops.workbench.dataset.DataSetServiceImpl;
+import org.pmiops.workbench.dataset.DataSetServiceImpl.QueryAndParameters;
 import org.pmiops.workbench.db.model.DbCohort;
 import org.pmiops.workbench.db.model.DbConceptSet;
 import org.pmiops.workbench.exceptions.BadRequestException;
